Rename README.md to README.md.
This clearly shows all the processes involved in analyzing this data on Power BI. The data is on "Personality Behavior". It clearly shows  step-by-step how I went on to source, clean, analyze and present this data on some of the behavioral patterns that distinguishes and identifies an introvert vs an extrovert in relation to their social behavior.